add printf parameters to PS2
This fixes the following compilation error in GCC 4.7.3:

    io.c: In function ‘ps2’:
    io.c:23:3: error: format not a string literal and no format arguments [-Werror=format-security]
    cc1: all warnings being treated as errors

GCC thinks that since PS2 is not a string literal, it might contain the
‘%n’ sequence, which puts the number of characters written into the
corresponding argument; since there are no arguments, the result would
be unpredictable and possibly dangerous.

This error does not appear in GCC 4.8; I assume they made GCC smarter so
that it sees that the PS2 variable is effectively a constant string that
doesn’t contain the ‘n’ conversion specifier. Anyways, by giving some
arguments to PS2 (the same as for PS1, plus the command being executed)
we avoid this compiler error.